Extract portions of FileDatastore.get()
Convert portions of FileDatastore.get() to free functions, to allow the get operation to be called on the client side without instantiating a FileDatastore object
